feat: export msgTransformer for other libraries
