+++ src/exported.ts
@@ -6,10 +6,11 @@
  */
 
 import { Messages } from './messages';
+
 Messages.importMessagesDirectory(__dirname);
 
 export { OAuth2Config } from 'jsforce';
-
+export { messageTransformer } from './messageTransformer';
 export { ConfigFile } from './config/configFile';
 
 export { TTLConfig } from './config/ttlConfig';
