+++ tsconfig.json
@@ -4,7 +4,7 @@
     "outDir": "./lib",
     "resolveJsonModule": true,
     "rootDir": "./src",
-    "plugins": [{ "transform": "./messageTransformer/messageTransformer.ts" }]
+    "plugins": [{ "transform": "./src/messageTransformer.ts" }]
   },
   "include": ["./src/**/*.ts"]
 }
